refactor(teststructure): drop aws, k8s, packer and ssh dependencies
teststructure required 7 of the 16 v2 modules, which made it the heaviest
module in the repo: importing it just for RunTestStage pulled in the AWS SDK
and client-go. That undercuts the point of splitting v2 into modules, since
teststructure is the entry point for most users.

Almost all of that weight came from typed Save/Load wrappers that existed only
to name a type in a signature. Those move to the module that owns the type:

  test_structure.SaveEc2KeyPair      -> aws.SaveEc2KeyPair
  test_structure.LoadEc2KeyPair      -> aws.LoadEc2KeyPair
  test_structure.SaveKubectlOptions  -> k8s.SaveKubectlOptions
  test_structure.LoadKubectlOptions  -> k8s.LoadKubectlOptions
  test_structure.SavePackerOptions   -> packer.SavePackerOptions
  test_structure.LoadPackerOptions   -> packer.LoadPackerOptions
  test_structure.SaveSSHKeyPair      -> ssh.SaveSSHKeyPair
  test_structure.LoadSSHKeyPair      -> ssh.LoadSSHKeyPair

Signatures, return types and on-disk filenames are unchanged, so callers only
need to change the package qualifier.

To let those modules share the save/load primitive without importing
teststructure, it moves to a new core/teststate package. The directory cannot
be named testdata because the go tool ignores directories with that name.
teststructure keeps its generic wrappers (SaveTestData, LoadTestData,
FormatTestDataPath, IsTestDataPresent, IsEmptyJSON, CleanupTestData*) as thin
delegations, so those callers are untouched.

teststructure keeps terraform and opa. They are what the package is for, and
they are light: 77 and 46 go.sum lines against 172 for aws and 259 for k8s.

teststate.SaveRedacted preserves the existing behaviour where an Ec2Keypair is
never written to the test log, along with the regression test covering it,
which moves to the aws module.

Result: teststructure go.sum drops from 301 to 104 lines, with no AWS SDK, no
client-go, and no transitive database drivers.

Refs #1875
