+++ modules/teststructure/save_test_data.go
@@ -1,39 +1,29 @@
 package teststructure
 
 import (
-	"encoding/json"
-	"os"
-	"path/filepath"
-
-	"github.com/gruntwork-io/terratest/modules/aws/v2"
-	"github.com/gruntwork-io/terratest/modules/core/v2/files"
-	"github.com/gruntwork-io/terratest/modules/core/v2/logger"
 	"github.com/gruntwork-io/terratest/modules/core/v2/testing"
-	"github.com/gruntwork-io/terratest/modules/k8s/v2"
-	"github.com/gruntwork-io/terratest/modules/packer/v2"
-	"github.com/gruntwork-io/terratest/modules/ssh/v2"
+	"github.com/gruntwork-io/terratest/modules/core/v2/teststate"
 	"github.com/gruntwork-io/terratest/modules/terraform/v2"
-	"github.com/stretchr/testify/require"
 )
 
 // SaveTerraformOptions serializes and saves TerraformOptions into the given folder. This allows you to create TerraformOptions during setup
 // and to reuse that TerraformOptions later during validation and teardown.
 func SaveTerraformOptions(t testing.TestingT, testFolder string, terraformOptions *terraform.Options) {
-	SaveTestData(t, formatTerraformOptionsPath(testFolder), true, terraformOptions)
+	teststate.Save(t, formatTerraformOptionsPath(testFolder), true, terraformOptions)
 }
 
 // SaveTerraformOptionsIfNotPresent serializes and saves TerraformOptions into the given folder if the file does not exist or the json is
 // empty. This allows you to create TerraformOptions during setup and to reuse that TerraformOptions later during validation and teardown,
 // but will prevent overwriting the contents and potentially duplicating resources.
 func SaveTerraformOptionsIfNotPresent(t testing.TestingT, testFolder string, terraformOptions *terraform.Options) {
-	SaveTestData(t, formatTerraformOptionsPath(testFolder), false, terraformOptions)
+	teststate.Save(t, formatTerraformOptionsPath(testFolder), false, terraformOptions)
 }
 
 // LoadTerraformOptions loads and unserializes TerraformOptions from the given folder. This allows you to reuse a TerraformOptions that was
 // created during an earlier setup step in later validation and teardown steps.
 func LoadTerraformOptions(t testing.TestingT, testFolder string) *terraform.Options {
 	var terraformOptions terraform.Options
-	LoadTestData(t, formatTerraformOptionsPath(testFolder), &terraformOptions)
+	teststate.Load(t, formatTerraformOptionsPath(testFolder), &terraformOptions)
 
 	return &terraformOptions
 }
@@ -43,98 +33,18 @@ func formatTerraformOptionsPath(testFolder string) string {
 	return FormatTestDataPath(testFolder, "TerraformOptions.json")
 }
 
-// SavePackerOptions serializes and saves PackerOptions into the given folder. This allows you to create PackerOptions during setup
-// and to reuse that PackerOptions later during validation and teardown.
-func SavePackerOptions(t testing.TestingT, testFolder string, packerOptions *packer.Options) {
-	SaveTestData(t, formatPackerOptionsPath(testFolder), true, packerOptions)
-}
-
-// LoadPackerOptions loads and unserializes PackerOptions from the given folder. This allows you to reuse a PackerOptions that was
-// created during an earlier setup step in later validation and teardown steps.
-func LoadPackerOptions(t testing.TestingT, testFolder string) *packer.Options {
-	var packerOptions packer.Options
-	LoadTestData(t, formatPackerOptionsPath(testFolder), &packerOptions)
-
-	return &packerOptions
-}
-
-// formatPackerOptionsPath formats a path to save PackerOptions in the given folder.
-func formatPackerOptionsPath(testFolder string) string {
-	return FormatTestDataPath(testFolder, "PackerOptions.json")
-}
-
-// SaveEc2KeyPair serializes and saves an Ec2KeyPair into the given folder. This allows you to create an Ec2KeyPair during setup
-// and to reuse that Ec2KeyPair later during validation and teardown.
-func SaveEc2KeyPair(t testing.TestingT, testFolder string, keyPair *aws.Ec2Keypair) {
-	saveTestData(t, formatEc2KeyPairPath(testFolder), true, keyPair, false)
-}
-
-// LoadEc2KeyPair loads and unserializes an Ec2KeyPair from the given folder. This allows you to reuse an Ec2KeyPair that was
-// created during an earlier setup step in later validation and teardown steps.
-func LoadEc2KeyPair(t testing.TestingT, testFolder string) *aws.Ec2Keypair {
-	var keyPair aws.Ec2Keypair
-	LoadTestData(t, formatEc2KeyPairPath(testFolder), &keyPair)
-
-	return &keyPair
-}
-
-// formatEc2KeyPairPath formats a path to save an Ec2KeyPair in the given folder.
-func formatEc2KeyPairPath(testFolder string) string {
-	return FormatTestDataPath(testFolder, "Ec2KeyPair.json")
-}
-
-// SaveSSHKeyPair serializes and saves an SSH key pair into the given folder. This allows you to create an SSH key pair
-// during setup and to reuse that key pair later during validation and teardown.
-func SaveSSHKeyPair(t testing.TestingT, testFolder string, keyPair *ssh.KeyPair) {
-	SaveTestData(t, formatSSHKeyPairPath(testFolder), true, keyPair)
-}
-
-// LoadSSHKeyPair loads and unserializes an SSH key pair from the given folder. This allows you to reuse an SSH key pair
-// that was created during an earlier setup step in later validation and teardown steps.
-func LoadSSHKeyPair(t testing.TestingT, testFolder string) *ssh.KeyPair {
-	var keyPair ssh.KeyPair
-	LoadTestData(t, formatSSHKeyPairPath(testFolder), &keyPair)
-
-	return &keyPair
-}
-
-// formatSSHKeyPairPath formats a path to save an SSH key pair in the given folder.
-func formatSSHKeyPairPath(testFolder string) string {
-	return FormatTestDataPath(testFolder, "SshKeyPair.json")
-}
-
-// SaveKubectlOptions serializes and saves KubectlOptions into the given folder. This allows you to create a KubectlOptions during setup
-// and reuse that KubectlOptions later during validation and teardown.
-func SaveKubectlOptions(t testing.TestingT, testFolder string, kubectlOptions *k8s.KubectlOptions) {
-	SaveTestData(t, formatKubectlOptionsPath(testFolder), true, kubectlOptions)
-}
-
-// LoadKubectlOptions loads and unserializes a KubectlOptions from the given folder. This allows you to reuse a KubectlOptions that was
-// created during an earlier setup step in later validation and teardown steps.
-func LoadKubectlOptions(t testing.TestingT, testFolder string) *k8s.KubectlOptions {
-	var kubectlOptions k8s.KubectlOptions
-	LoadTestData(t, formatKubectlOptionsPath(testFolder), &kubectlOptions)
-
-	return &kubectlOptions
-}
-
-// formatKubectlOptionsPath formats a path to save a KubectlOptions in the given folder.
-func formatKubectlOptionsPath(testFolder string) string {
-	return FormatTestDataPath(testFolder, "KubectlOptions.json")
-}
-
 // SaveString serializes and saves a uniquely named string value into the given folder. This allows you to create one or more string
 // values during one stage -- each with a unique name -- and to reuse those values during later stages.
 func SaveString(t testing.TestingT, testFolder string, name string, val string) {
 	path := formatNamedTestDataPath(testFolder, name)
-	SaveTestData(t, path, true, val)
+	teststate.Save(t, path, true, val)
 }
 
 // LoadString loads and unserializes a uniquely named string value from the given folder. This allows you to reuse one or more string
 // values that were created during an earlier setup step in later steps.
 func LoadString(t testing.TestingT, testFolder string, name string) string {
 	var val string
-	LoadTestData(t, formatNamedTestDataPath(testFolder, name), &val)
+	teststate.Load(t, formatNamedTestDataPath(testFolder, name), &val)
 
 	return val
 }
@@ -143,14 +53,14 @@ func LoadString(t testing.TestingT, testFolder string, name string) string {
 // values during one stage -- each with a unique name -- and to reuse those values during later stages.
 func SaveInt(t testing.TestingT, testFolder string, name string, val int) {
 	path := formatNamedTestDataPath(testFolder, name)
-	SaveTestData(t, path, true, val)
+	teststate.Save(t, path, true, val)
 }
 
 // LoadInt loads a uniquely named int value from the given folder. This allows you to reuse one or more int
 // values that were created during an earlier setup step in later steps.
 func LoadInt(t testing.TestingT, testFolder string, name string) int {
 	var val int
-	LoadTestData(t, formatNamedTestDataPath(testFolder, name), &val)
+	teststate.Load(t, formatNamedTestDataPath(testFolder, name), &val)
 
 	return val
 }
@@ -176,7 +86,7 @@ func formatNamedTestDataPath(testFolder string, name string) string {
 
 // FormatTestDataPath formats a path to save test data.
 func FormatTestDataPath(testFolder string, filename string) string {
-	return filepath.Join(testFolder, ".test-data", filename)
+	return teststate.FormatPath(testFolder, filename)
 }
 
 // SaveTestData serializes and saves a value used at test time to the given path. This allows you to create some sort of test data
@@ -184,173 +94,39 @@ func FormatTestDataPath(testFolder string, filename string) string {
 // any contents that exist in the file found at `path` will be overwritten. This has the potential for causing duplicated resources
 // and should be used with caution. If `overwrite` is `false`, the save will be skipped and a warning will be logged.
 func SaveTestData(t testing.TestingT, path string, overwrite bool, value any) {
-	saveTestData(t, path, overwrite, value, true)
-}
-
-// saveTestData serializes and saves a value used at test time to the given path. This allows you to create some sort of test data
-// (e.g., TerraformOptions) during setup and to reuse this data later during validation and teardown. If `overwrite` is `true`,
-// any contents that exist in the file found at `path` will be overwritten. This has the potential for causing duplicated resources
-// and should be used with caution. If `overwrite` is `false`, the save will be skipped and a warning will be logged.
-// If `loggedVal` is `true`, the value will be logged as JSON.
-func saveTestData(t testing.TestingT, path string, overwrite bool, value any, loggedVal bool) {
-	logger.Default.Logf(t, "Storing test data in %s so it can be reused later", path)
-
-	if IsTestDataPresent(t, path) {
-		if overwrite {
-			logger.Default.Logf(t, "[WARNING] The named test data at path %s is non-empty. Save operation will overwrite existing value with \"%v\".\n.", path, value)
-		} else {
-			logger.Default.Logf(t, "[WARNING] The named test data at path %s is non-empty. Skipping save operation to prevent overwriting existing value with \"%v\".\n.", path, value)
-
-			return
-		}
-	}
-
-	bytes, err := json.Marshal(value)
-	if err != nil {
-		t.Fatalf("Failed to convert value %s to JSON: %v", path, err)
-	}
-
-	if loggedVal {
-		logger.Default.Logf(t, "Marshalled JSON: %s", string(bytes))
-	}
-
-	parentDir := filepath.Dir(path)
-
-	if err := os.MkdirAll(parentDir, 0o755); err != nil {
-		t.Fatalf("Failed to create folder %s: %v", parentDir, err)
-	}
-
-	if err := os.WriteFile(path, bytes, 0o644); err != nil {
-		t.Fatalf("Failed to save value %s: %v", path, err)
-	}
+	teststate.Save(t, path, overwrite, value)
 }
 
 // LoadTestData loads and unserializes a value stored at the given path. The value should be a pointer to a struct into which the
 // value will be deserialized. This allows you to reuse some sort of test data (e.g., TerraformOptions) from earlier
 // setup steps in later validation and teardown steps.
 func LoadTestData(t testing.TestingT, path string, value any) {
-	logger.Default.Logf(t, "Loading test data from %s", path)
-
-	bytes, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatalf("Failed to load value from %s: %v", path, err)
-	}
-
-	if err := json.Unmarshal(bytes, value); err != nil {
-		t.Fatalf("Failed to parse JSON for value %s: %v", path, err)
-	}
+	teststate.Load(t, path, value)
 }
 
 // IsTestDataPresent returns true if a file exists at $path and the test data there is non-empty.
 func IsTestDataPresent(t testing.TestingT, path string) bool {
-	exists, err := files.FileExistsE(path)
-	if err != nil {
-		t.Fatalf("Failed to load test data from %s due to unexpected error: %v", path, err)
-	}
-
-	if !exists {
-		return false
-	}
-
-	bytes, err := os.ReadFile(path)
-	if err != nil {
-		t.Fatalf("Failed to load test data from %s due to unexpected error: %v", path, err)
-	}
-
-	if IsEmptyJSON(t, bytes) {
-		return false
-	}
-
-	return true
+	return teststate.IsPresent(t, path)
 }
 
 // IsEmptyJSON returns true if the given bytes are empty, or in a valid JSON format that can reasonably be considered empty.
 // The types used are based on the type possibilities listed at https://golang.org/src/encoding/json/decode.go?s=4062:4110#L51
 func IsEmptyJSON(t testing.TestingT, bytes []byte) bool {
-	var value any
-
-	if len(bytes) == 0 {
-		return true
-	}
-
-	if err := json.Unmarshal(bytes, &value); err != nil {
-		t.Fatalf("Failed to parse JSON while testing whether it is empty: %v", err)
-	}
-
-	if value == nil {
-		return true
-	}
-
-	valueBool, ok := value.(bool)
-	if ok && !valueBool {
-		return true
-	}
-
-	valueFloat64, ok := value.(float64)
-	if ok && valueFloat64 == 0 {
-		return true
-	}
-
-	valueString, ok := value.(string)
-	if ok && valueString == "" {
-		return true
-	}
-
-	valueSlice, ok := value.([]any)
-	if ok && len(valueSlice) == 0 {
-		return true
-	}
-
-	valueMap, ok := value.(map[string]any)
-	if ok && len(valueMap) == 0 {
-		return true
-	}
-
-	return false
+	return teststate.IsEmptyJSON(t, bytes)
 }
 
 // CleanupTestData cleans up the test data at the given path.
 func CleanupTestData(t testing.TestingT, path string) {
-	if files.FileExists(path) {
-		logger.Default.Logf(t, "Cleaning up test data from %s", path)
-
-		if err := os.Remove(path); err != nil {
-			t.Fatalf("Failed to clean up file at %s: %v", path, err)
-		}
-	} else {
-		logger.Default.Logf(t, "%s does not exist. Nothing to cleanup.", path)
-	}
+	teststate.Cleanup(t, path)
 }
 
 // CleanupTestDataFolder cleans up the .test-data folder inside the given folder.
 // If there are any errors, fail the test.
 func CleanupTestDataFolder(t testing.TestingT, path string) {
-	err := CleanupTestDataFolderE(t, path)
-	require.NoError(t, err)
+	teststate.CleanupFolder(t, path)
 }
 
 // CleanupTestDataFolderE cleans up the .test-data folder inside the given folder.
 func CleanupTestDataFolderE(t testing.TestingT, path string) error {
-	path = filepath.Join(path, ".test-data")
-
-	exists, err := files.FileExistsE(path)
-	if err != nil {
-		logger.Default.Logf(t, "Failed to clean up test data folder at %s: %v", path, err)
-
-		return err
-	}
-
-	if !exists {
-		logger.Default.Logf(t, "%s does not exist. Nothing to cleanup.", path)
-
-		return nil
-	}
-
-	if err := os.RemoveAll(path); err != nil {
-		logger.Default.Logf(t, "Failed to clean up test data folder at %s: %v", path, err)
-
-		return err
-	}
-
-	return nil
+	return teststate.CleanupFolderE(t, path)
 }
